requirements.txt
fastapi==0.110.1
uvicorn==0.30.1
pydantic==2.6.3
langchain==0.1.4
langchain-community==0.0.19
langchain-openai==0.0.5
python-dotenv==1.0.1
sentence-transformers==2.5.0
detoxify==0.5.2
transformers==4.38.1
torch
torchvision
torchaudio
pydantic-settings==2.1.0
openai==1.14.2
datasets
accelerate
evaluate
peft
bitsandbytes
sentencepiece
scikit-learn
pandas
tqdm
wandb 